clean up ruby tests
