--- tests/rspec/spec/aws/custom_tls_spec.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-require 'shared_examples/k8s'
-
-DOMAIN = 'tectonic-ci.de'
-
-RSpec.describe 'aws-custom-tls' do
-  include_examples('withBuildFolderSetup', File.join(ENV['RSPEC_PATH'], '../smoke/aws/vars/aws.tfvars.json'))
-  include_examples('withTLSSetup', DOMAIN)
-
-  before(:all) do
-    # doing this because the example aws.tfvars.json is being used by the UI tests
-    # and it generate a fix cluster name. Here we just change the cluster/dns name
-    # all the other content stay the same as generated by the UI
-    # If we use the same name generated by the UI we can run just one test at the time.
-    @tfvars_file.change_cluster_name(@name)
-    @tfvars_file.change_dns_name(@name)
-  end
-
-  context 'with a cluster' do
-    include_examples('withRunningClusterExistingBuildFolder', true)
-  end
-end
