--- tests/smoke/aws/vars/aws-basic.tfvars.json
@@ -1,21 +1,0 @@
-{
-  "tectonic_aws_etcd_ec2_type": "m4.large",
-  "tectonic_aws_etcd_root_volume_size": 32,
-  "tectonic_aws_etcd_root_volume_type": "gp2",
-  "tectonic_aws_master_ec2_type": "m4.large",
-  "tectonic_aws_master_root_volume_size": 32,
-  "tectonic_aws_master_root_volume_type": "gp2",
-  "tectonic_aws_endpoints": "public",
-  "tectonic_aws_vpc_cidr_block": "10.0.0.0/16",
-  "tectonic_aws_worker_ec2_type": "m4.large",
-  "tectonic_aws_worker_root_volume_size": 32,
-  "tectonic_aws_worker_root_volume_type": "gp2",
-  "tectonic_base_domain": "tectonic-ci.de",
-  "tectonic_cluster_name": "aws-basic",
-  "tectonic_dns_name": "aws-basic",
-  "tectonic_etcd_count": 3,
-  "tectonic_master_count": 2,
-  "tectonic_stats_url": "https://stats-collector-staging.tectonic.com",
-  "tectonic_worker_count": 2,
-  "tectonic_container_linux_channel": "beta"
-}
