--- tests/smoke/aws/vars/aws-exp.tfvars.json
@@ -1,25 +1,0 @@
-{
-    "tectonic_worker_count": 2,
-
-    "tectonic_master_count": 1,
-
-    "tectonic_ca_cert": "",
-
-    "tectonic_ca_key": "",
-
-    "tectonic_aws_master_ec2_type": "m4.large",
-
-    "tectonic_aws_worker_ec2_type": "m4.large",
-
-    "tectonic_aws_etcd_ec2_type": "m4.large",
-
-    "tectonic_aws_vpc_cidr_block": "10.0.0.0/16",
-
-    "tectonic_aws_external_vpc_id": "",
-
-    "tectonic_experimental": true,
-
-    "tectonic_stats_url": "https://stats-collector-staging.tectonic.com",
-
-    "tectonic_container_linux_channel": "beta"
-}
