--- tests/smoke/aws/vars/aws-exp.yaml
@@ -1,22 +1,0 @@
-aws:
-  etcd:
-    ec2Type: m4.large
-  master:
-    ec2Type: m4.large
-  vpcCIDRBlock: 10.0.0.0/16
-  worker:
-    ec2Type: m4.large
-containerLinux:
-  channel: beta
-master:
-  nodepools:
-    - master
-nodePools:
-  - count: 1
-    name: master
-  - count: 2
-    name: worker
-platform: aws
-worker:
-  nodePools:
-    - worker
