+++ tests/smoke/cluster_test.go
@@ -77,7 +77,8 @@ func testCluster(t *testing.T) {
 	//	t.Run("NetworkPolicy", testNetworkPolicy)
 	//}
 
-	t.Run("KillAPIServer", testKillAPIServer)
+	// TODO: temporary disabling this for OpenTonic
+	//t.Run("KillAPIServer", testKillAPIServer)
 }
 
 func testAllPodsRunning(t *testing.T) {
